+++ src/Components/Firebase/context.js
@@ -1,0 +1,5 @@
+import React from 'react';
+
+const FirebaseContext = React.createContext(null);
+
+export default FirebaseContext;
