+++ src/Components/Firebase/index.js
@@ -1,0 +1,4 @@
+import FirebaseContext from './context';
+import Firebase from './firebase';
+export default Firebase;
+export { FirebaseContext };
